+++ src/features/create-post/ui/PublicationStep.tsx
@@ -8,7 +8,7 @@ import {
   normalizeCreatePostDescription,
 } from '../model/createPostDescription'
 import type { CreatePostPhoto } from '../model/createPostFile'
-import styles from './createPostPage.module.css'
+import styles from './createPost.module.css'
 import { EditedPhotoPreview } from './EditedPhotoPreview'
 import { PhotoCarouselControls } from './PhotoCarouselControls'